Move integrations tests
